fix(32-bit): refuse an unrepresentable cache budget, and gate the class in CI
cacheGeometry returned int(perShard) from an int64 with no bound. On a 32-bit
build a per-shard share above MaxInt wrapped silently — 32 GiB truncates to 0 —
and the failure surfaced later out of Config.Validate, naming the wrong cause.
It now refuses with the real reason.

Three test files also held untyped constants that do not fit a 32-bit int, which
is a COMPILE error rather than a runtime one, so those packages could not be
built for 386 at all:

  - cachebudget_test.go carried 32 * GiB as an int expectation; the want field is
    now int64 and the case asserts the refusal above where it cannot be
    represented.
  - semcache/embedder_test.go and inttest/mv_scroll_cursor_test.go both used the
    golden-ratio constant 0x9E3779B1 in int arithmetic. Both now do the
    multiplication in an unsigned type, which is byte-identical on 64-bit: the
    low 32 bits of a product do not depend on the high bits.

The CI gate is what keeps this from recurring. No lane compiled test files under
GOARCH=386 — the 32-bit lane skips the packages whose budgets cannot fit a
32-bit address space, and the cross-compile matrix runs `go build`, which ignores
_test.go entirely. `go vet ./...` type-checks every package including its tests
without building a binary or allocating, so it covers what the test lane cannot.

+++ inttest/mv_scroll_cursor_test.go
@@ -208,7 +208,7 @@ func TestMVScrollPaginationStableAcrossPartitions(t *testing.T) {
 		}
 		// Shuffle the combined seed order so core/volatile interleave across partitions.
 		for i := range seedIDs {
-			j := (i*2654435761 + 12345) % len(seedIDs)
+			j := int((uint64(i)*2654435761 + 12345) % uint64(len(seedIDs)))
 			seedIDs[i], seedIDs[j] = seedIDs[j], seedIDs[i]
 		}
 		for _, id := range seedIDs {
